pyproject.toml
[tool.poetry]
name = "sample"
version = "0.1.0"
description = ""
authors = ""
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "2.114.1"
constructs = ">=10.0.0,<11.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "6.2.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
